requirements.txt
streamlit
pandas
SQLAlchemy
psycopg2-binary
python-dotenv
plotly
google-genai
sqlparse
